Add use group-focus-within on whole component
Previously, conditional rendering such as highlighting the input
box, showing the left hand side magnifying glass and showing the
suggestions dropdown was done using the state of the input
container (3rd div).

However, this did not give a way to keep the focussed behaviour
(in particular making the suggestions dropdown visible) active
when clicking anywhere in the search bar outside of the input box
itself. This meant the dropdown would disappear before it's button
clicks would get activated.

Instead, we move all focus logic to the components parent div and
reference it as a group property. Thus if we click a dropdown button,
it will be focussed, and thus the group-focus-within will still be
true. The dropdown will remain rendered and the onClick function
will be called.

Note that to ease usability, we return the focus to the input element
on all callbacks. I.E when we accept a suggestion, we return focus to
the input so that the user can keep on typing.